requirements.txt
mysql-connector-python~=9.0.0
fastapi~=0.115.0
